static cast

+++ oop2_ex4/GameController.cpp
@@ -143,7 +143,7 @@ void GameController::playGame(Timer& screenUpdatesTimer, GameScreen& gameScreen,
 				
 				// update area percent
 				std::shared_ptr<AreaButton>& myAreaBT = gameScreen.getBottomPanel()->getMyAreaButton();
-				myAreaBT->updateAreaPercent(userPlayer->getPlayerVertices().size(), gameScreen.getBoard()->getPolygonsGraph().getNumOfVertices());
+				myAreaBT->updateAreaPercent(static_cast<int>(userPlayer->getPlayerVertices().size()), static_cast<int>(gameScreen.getBoard()->getPolygonsGraph().getNumOfVertices()));
 				myAreaBT->setText(myAreaBT->getPreText() + std::to_string(myAreaBT->getAreaPercent()) + "%");
 				if (myAreaBT->getAreaPercent() >= 20.f) {
 					gameScreen.close();
@@ -166,7 +166,7 @@ void GameController::playGame(Timer& screenUpdatesTimer, GameScreen& gameScreen,
 					
 					// update area percent
 					std::shared_ptr<AreaButton>& rivalAreaBT = gameScreen.getBottomPanel()->getRivalAreaButton();
-					rivalAreaBT->updateAreaPercent(otherPlayer->getPlayerVertices().size(), gameScreen.getBoard()->getPolygonsGraph().getNumOfVertices());
+					rivalAreaBT->updateAreaPercent(static_cast<int>(otherPlayer->getPlayerVertices().size()), static_cast<int>(gameScreen.getBoard()->getPolygonsGraph().getNumOfVertices()));
 					rivalAreaBT->setText(rivalAreaBT->getPreText() + std::to_string(rivalAreaBT->getAreaPercent()) + "%");
 				}
 			}
